requirements.txt
beautifulsoup4==4.12.3
jieba==0.42.1
lxml==5.2.2
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
Requests==2.32.3
wordcloud==1.9.3
